+++ src/include/execution/executors/topn_executor.h
@@ -13,6 +13,8 @@
 #pragma once
 
 #include <cstddef>
+#include <deque>
+#include <functional>
 #include <memory>
 #include <queue>
 #include <utility>
@@ -25,6 +27,70 @@
 #include "execution/plans/seq_scan_plan.h"
 #include "execution/plans/topn_plan.h"
 #include "storage/table/tuple.h"
+
+namespace bustub {
+struct HeapKeyType {
+  HeapKeyType(Tuple tuple, const std::vector<std::pair<OrderByType, AbstractExpressionRef>> &order_by,
+              AbstractExecutor *child_executor)
+      : tuple_(std::move(tuple)), order_by_(order_by), child_executor_(child_executor) {}
+
+  HeapKeyType(const HeapKeyType &rhs) : order_by_(rhs.order_by_) {
+    tuple_ = rhs.tuple_;
+    child_executor_ = rhs.child_executor_;
+  }
+
+  HeapKeyType(HeapKeyType &&rhs) noexcept : order_by_(rhs.order_by_) {
+    tuple_ = std::move(rhs.tuple_);
+    child_executor_ = rhs.child_executor_;
+  }
+
+  auto operator=(const HeapKeyType &rhs) -> HeapKeyType & {
+    tuple_ = rhs.tuple_;
+    return *this;
+  }
+
+  auto operator=(HeapKeyType &&rhs) noexcept -> HeapKeyType & {
+    tuple_ = std::move(rhs.tuple_);
+    return *this;
+  }
+
+  Tuple tuple_;
+  const std::vector<std::pair<OrderByType, AbstractExpressionRef>> &order_by_;
+  AbstractExecutor *child_executor_;
+};
+
+}  // namespace bustub
+
+namespace std {
+template <>
+struct less<bustub::HeapKeyType> {
+  auto operator()(const bustub::HeapKeyType &lhs, const bustub::HeapKeyType &rhs) const -> bool {
+    auto child_executor = lhs.child_executor_;
+    for (auto &p : lhs.order_by_) {
+      auto order = p.first;
+      auto &exp = p.second;
+      auto lvalue = exp->Evaluate(&lhs.tuple_, child_executor->GetOutputSchema());
+      auto rvalue = exp->Evaluate(&rhs.tuple_, child_executor->GetOutputSchema());
+      if (order == bustub::OrderByType::DESC) {
+        if (lvalue.CompareGreaterThan(rvalue) == bustub::CmpBool::CmpTrue) {
+          return true;
+        }
+        if (lvalue.CompareLessThan(rvalue) == bustub::CmpBool::CmpTrue) {
+          return false;
+        }
+      } else {
+        if (lvalue.CompareLessThan(rvalue) == bustub::CmpBool::CmpTrue) {
+          return true;
+        }
+        if (lvalue.CompareGreaterThan(rvalue) == bustub::CmpBool::CmpTrue) {
+          return false;
+        }
+      }
+    }
+    UNREACHABLE("duplicate key is not allowed");
+  }
+};
+}  // namespace std
 
 namespace bustub {
